ah crap i did it again
well ive completed everything including the stretch except for animation. So ill explain some of it. For the cards i did some crazy stuff i used the for/in as the response.data.articles isnt enumerable this returns the keys of the articls thehn i use that to do a foreach loop for the articles and set the card data that way. The other ones were really simple so ill explain carousel in this one i create a const array imgs and add all the img elements to it i also set an index of 0. Then when creating the caraousel i set the first image as visible. in the move function i set a variabled called previndex before changing index this is so that i can edit the value of the previous img display property and the current. oh and i check which button triggered the event and do the proper math accordingly.